layer tests: Consolidate framework files
