+++ layer/tests/profiles_test_helper.h
@@ -25,12 +25,7 @@
 #include <vulkan/vulkan.h>
 #include "../profiles.h"
 #include "gtest/gtest.h"
-
-#if defined(ANDROID)
-#include "vktestframeworkandroid.h"
-#else
 #include "vktestframework.h"
-#endif
 
 namespace profiles_test {
