--- layer/tests/vktestframeworkandroid.cpp
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2015-2022 The Khronos Group Inc.
- * Copyright (c) 2015-2022 Valve Corporation
- * Copyright (c) 2015-2022 LunarG, Inc.
- * Copyright (c) 2015-2022 Google, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Author: Mark Lobodzinski <Mark@LunarG.com>
- */
-
-#include "vktestframeworkandroid.h"
-#include <android/log.h>
-
-VkTestFramework::VkTestFramework() {}
-VkTestFramework::~VkTestFramework() {}
-
-// Define static elements
-ANativeWindow *VkTestFramework::window = nullptr;
